+++ src/triangle_mesh.cpp
@@ -39,7 +39,7 @@ public:
         } else {
             // If binary is not found, load obj then export binary.
             auto vertices = std::vector<edyn::vector3>{};
-            auto indices = std::vector<uint16_t>{};
+            auto indices = std::vector<uint32_t>{};
             auto scale = edyn::scalar(0.1) * edyn::vector3_one;
             edyn::load_tri_mesh_from_obj("../../../edyn-testbed/resources/terrain.obj",
                                          vertices, indices, nullptr, {0,0,0}, {0,0,0,1}, scale);
